+++ widgets/client/messenger/components/websiteApp/WebsiteAppDetail.tsx
@@ -1,29 +1,34 @@
-import * as React from "react";
-import { iconLeft } from "../../../icons/Icons";
-import { __ } from "../../../utils";
-import TopBar from "../../containers/TopBar";
-import { IWebsiteApp } from "../../types";
+import * as React from 'react';
+import { iconLeft } from '../../../icons/Icons';
+import { __ } from '../../../utils';
+import TopBar from '../../containers/TopBar';
+import { IWebsiteApp } from '../../types';
 
 type Props = {
   websiteApp: IWebsiteApp;
   changeRoute: (route: string) => void;
+  loading: boolean;
 };
 
 export default class WebsiteAppDetail extends React.PureComponent<Props> {
   render() {
-    const { changeRoute, websiteApp } = this.props;
+    const { changeRoute, websiteApp, loading } = this.props;
 
-    const onClick = () => changeRoute("home");
+    const onClick = () => changeRoute('home');
 
     return (
       <>
         <TopBar
-          middle={websiteApp.credentials.description}
+          middle={loading ? '' : websiteApp.credentials.description}
           buttonIcon={iconLeft()}
           onLeftButtonClick={onClick}
         />
         <div className="erxes-content">
-          <iframe src={websiteApp.credentials.url} className="websiteApp" />
+          {loading ? (
+            <div className="loader" />
+          ) : (
+            <iframe src={websiteApp.credentials.url} className="websiteApp" />
+          )}
         </div>
       </>
     );